pyproject.toml
[project]
name = "kite-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.13",
    "asyncio>=3.4.3",
    "browser-cookie3>=0.20.1",
    "fastapi>=0.115.13",
    "fastapi-users-db-sqlalchemy>=7.0.0",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "fastmcp>=2.5.1",
    "jinja2>=3.1.6",
    "newsapi>=0.1.1",
    "newsapi-python>=0.2.7",
    "openai>=1.90.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "yfinance>=0.2.63",
]
